+++ packages/playground/vue-sourcemap/Js.vue
@@ -1,0 +1,11 @@
+<template>
+  <p>&lt;js&gt;</p>
+</template>
+
+<script>
+console.log('script')
+</script>
+
+<script setup>
+console.log('setup')
+</script>