+++ packages/playground/worker/modules/module3.js
@@ -1,0 +1,2 @@
+export * from './module'
+export const msg3 = 'module3'