+++ packages/playground/worker/url-shared-worker.js
@@ -1,4 +1,4 @@
-import constant from './module'
+import constant from './modules/module'
 
 self.onconnect = (event) => {
   const port = event.ports[0]