+++ packages/vite/src/node/plugins/importAnalysisBuild.ts
@@ -86,6 +86,7 @@ function preload(baseModule: () => Promise<{}>, deps?: string[]) {
 export function buildImportAnalysisPlugin(config: ResolvedConfig): Plugin {
   const ssr = !!config.build.ssr
   const insertPreload = !(ssr || !!config.build.lib)
+  const isWorker = config.isWorker
 
   const scriptRel = config.build.polyfillModulePreload
     ? `'modulepreload'`
@@ -120,6 +121,11 @@ export function buildImportAnalysisPlugin(config: ResolvedConfig): Plugin {
         return
       }
 
+      if (isWorker) {
+        // preload method use `document` and can't run in the worker
+        return
+      }
+
       await init
 
       let imports: readonly ImportSpecifier[] = []
@@ -132,7 +138,6 @@ export function buildImportAnalysisPlugin(config: ResolvedConfig): Plugin {
       if (!imports.length) {
         return null
       }
-
       let s: MagicString | undefined
       const str = () => s || (s = new MagicString(source))
       let needPreloadHelper = false
@@ -165,7 +170,7 @@ export function buildImportAnalysisPlugin(config: ResolvedConfig): Plugin {
               insertPreload
             )
           str().prepend(importsString)
-          str().overwrite(expStart, endIndex, exp)
+          str().overwrite(expStart, endIndex, exp, { contentOnly: true })
           if (!isEager) {
             needPreloadHelper = true
           }
@@ -176,7 +181,7 @@ export function buildImportAnalysisPlugin(config: ResolvedConfig): Plugin {
           needPreloadHelper = true
           const original = source.slice(expStart, expEnd)
           const replacement = `${preloadMethod}(() => ${original},${isModernFlag}?"${preloadMarker}":void 0)`
-          str().overwrite(expStart, expEnd, replacement)
+          str().overwrite(expStart, expEnd, replacement, { contentOnly: true })
         }
 
         // Differentiate CSS imports that use the default export from those that
@@ -191,7 +196,9 @@ export function buildImportAnalysisPlugin(config: ResolvedConfig): Plugin {
           !(bareImportRE.test(specifier) && !specifier.includes('/'))
         ) {
           const url = specifier.replace(/\?|$/, (m) => `?used${m ? '&' : ''}`)
-          str().overwrite(start, end, dynamicIndex > -1 ? `'${url}'` : url)
+          str().overwrite(start, end, dynamicIndex > -1 ? `'${url}'` : url, {
+            contentOnly: true
+          })
         }
       }
 
@@ -223,7 +230,8 @@ export function buildImportAnalysisPlugin(config: ResolvedConfig): Plugin {
             s.overwrite(
               match.index,
               match.index + isModernFlag.length,
-              isModern
+              isModern,
+              { contentOnly: true }
             )
           }
           return {
@@ -238,7 +246,7 @@ export function buildImportAnalysisPlugin(config: ResolvedConfig): Plugin {
     },
 
     generateBundle({ format }, bundle) {
-      if (format !== 'es' || ssr) {
+      if (format !== 'es' || ssr || isWorker) {
         return
       }
 
@@ -303,7 +311,9 @@ export function buildImportAnalysisPlugin(config: ResolvedConfig): Plugin {
                         hasRemovedPureCssChunk = true
                       }
 
-                      s.overwrite(expStart, expEnd, 'Promise.resolve({})')
+                      s.overwrite(expStart, expEnd, 'Promise.resolve({})', {
+                        contentOnly: true
+                      })
                     }
                   }
                 }
@@ -330,7 +340,8 @@ export function buildImportAnalysisPlugin(config: ResolvedConfig): Plugin {
                     // main chunk is removed
                     (hasRemovedPureCssChunk && deps.size > 0)
                     ? `[${[...deps].map((d) => JSON.stringify(d)).join(',')}]`
-                    : `[]`
+                    : `[]`,
+                  { contentOnly: true }
                 )
               }
             }